pi-tomemd-miguelcc/requirements.txt
Django>=3.0,<4.0
django-crispy-forms
pandas
pendulum
requests
matplotlib